Rename id3c.db.cli to id3c.cli
The CLI started life in the at-the-time-unnamed database project.  Since
then it has grown to encompass commands not just limited to the
database.  It seems appropriate for the top-level id3c package to
contain api, db, and (now) cli subpackages.
