+++ lib/id3c/cli/__main__.py
@@ -1,0 +1,7 @@
+"""
+Invoked via ``python -m id3c.cli``.
+"""
+
+if __name__ == "__main__":
+    from id3c.cli import cli
+    cli(prog_name = "id3c")
