+++ lib/id3c/cli/command/clinical.py
@@ -15,7 +15,7 @@ import pandas as pd
 import id3c.db as db
 from math import ceil
 from id3c.db.session import DatabaseSession
-from id3c.db.cli import cli
+from id3c.cli import cli
 from . import (
     add_metadata,
     barcode_quality_control,
