+++ lib/id3c/cli/command/consensus_genome.py
@@ -4,7 +4,7 @@ Backfill previously generated consensus genomes that were not uploaded to ID3C.
 import click
 import logging
 import json
-from id3c.db.cli import cli
+from id3c.cli import cli
 from id3c.db.session import DatabaseSession
