+++ lib/id3c/cli/command/etl/manifest.py
