+++ lib/id3c/cli/command/longitudinal.py
@@ -15,7 +15,7 @@ import id3c.db as db
 from typing import List
 from itertools import combinations
 from id3c.db.session import DatabaseSession
-from id3c.db.cli import cli
+from id3c.cli import cli
 from . import (
     add_metadata,
     barcode_quality_control,
