+++ lib/id3c/cli/command/user.py
@@ -7,7 +7,7 @@ import id3c.db as db
 from operator import attrgetter
 from typing import List
 from id3c.db.session import DatabaseSession
-from id3c.db.cli import cli
+from id3c.cli import cli
 
 
 LOG = logging.getLogger(__name__)
